+++ src/architectures/base_model.py
