+++ src/core/base_conv/base_conv.py
@@ -1,0 +1,25 @@
+from abc import ABC, abstractmethod
+from typing import *
+import math
+from functools import partial
+from typing import Dict, Any
+import numpy as np
+import torch
+
+class BaseConvolution(ABC, torch.nn.Module):
+    def __init__(self, sampler, neighbour_finder, *args, **kwargs):
+        torch.nn.Module.__init__(self)
+
+        self.sampler = sampler
+        self.neighbour_finder = neighbour_finder
+
+    @property
+    def nb_params(self):
+        """[This property is used to return the number of trainable parameters for a given layer]
+        It is useful for debugging and reproducibility.
+        Returns:
+            [type] -- [description]
+        """
+        model_parameters = filter(lambda p: p.requires_grad, self.parameters())
+        self._nb_params = sum([np.prod(p.size()) for p in model_parameters])
+        return self._nb_params
